The gui-branch
It is probably better to implement the GTK+ GUI in a separate branch because some of the behaviour of the main script achtelbass.py must be changed. For instance the parameters that the GUI gives to the main script must be read from stdin and so. Changes like that in the main script can have unsuspected side effects and are better outsourced to a separate branch.