Experiments with Menubar.
Menubar is for "File -> Open", "File -> Save as" or "File -> Quit" and so.
This is necessary too, but what is most needed for now is rather
gtk.ComboBox: A kind of Button that expands downwards when klicked, showing
all the alternatives to chose. I hope at least that thats the right thing.