Added a locales file
The locales files contains a dictionary for all the strings that are
displayed on the buttons and so.
Translations to other languages can be added easily by copying the
locales_template.py file and replacing the values in the dictionary with
the terms in the target language. Then the new locales file must be
imported in the gachtelbass file.

Started to fix the incompatibilities of the strings which gachtelbass
passes to achtelbass which expects floats. It makes a messy impression on
me. I don't like too much to map the strings to floats in the __init__ of
achtelbass. Maybe it can be done in the generate method in gachtelbass.