Completed conversion of params from str to float
As far as i know the values in the dropdown lists in GTK+ are strings only.
But some of the params achtelbass needs are floats, e.g. note values. Those
need to be converted from strings to floats. This is done in the __init__
method of achtelbass. I decided to keep it there, so on the command line it
will be possible to pass human readable params like 1/8 instead of 0.125
for the note value.