Fix the problem of never reaching min_pitch
Fix the problem that in pitches.py neither min_pitch nor max_pitch were
never chosen, thus the span between min_pitch and max_pitch was too small.

Add a checking whether the tonic note is in the selectable notes and if so
then select the lowest selectable tonic as the first note to print out.