Add checking for bad time signature and note value
When time signature is 3/4 and the only note value is 1 or 1/2, program
crashes. Ungood. Add a warningdialog in GUI. Add an error in CLI.

Besides, fix the error in note_values.py that removing those note_values
from selectable_note_values list when the remaining bar_length get to
short vor those note_values was done after picking a note_value from
that list. The two statements must be swapped, then it works.