Fix that only crosses were set by changing key
Fix that mostly crosses, seldom and then only one flat were set when
changing key is chosen. That is because old_accidentals in
get_new_accidentals() was passed as a string, not as an int.
Furthermore, replace the excessive regex use by if clauses and list
checking. Needs to be continued though.